requirements.txt
crewai>=0.11.0
gitpython>=3.1.40
slack-sdk>=3.26.1
python-dotenv>=1.0.0
schedule>=1.2.1
langchain>=0.1.0
openai>=1.10.0
langchain-openai>=0.0.5
langchain-anthropic>=0.1.1
watchdog>=3.0.0
flask>=2.3.0
pyyaml>=6.0.0
